pyproject.toml
[tool.poetry]
name = "movierater"
version = "0.1.0"
description = ""
authors = ["Artem Shevelyukhin <artem_16_97@mail.ru>"]
readme = "README.md"
package-mode = false


[tool.poetry.dependencies]
python = "^3.14"
sqlalchemy = "^2.0.45"
fastapi = "^0.127.0"
python-dotenv = "^1.2.1"
nanoid = "^2.0.0"
uvicorn = "^0.40.0"
httpx = "^0.28.1"
jinja2 = "^3.1.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
